move `Environment` and `InEnvironment` into `ir`
