+++ chalk-rust/src/fold/mod.rs
@@ -1,6 +1,5 @@
 use errors::*;
 use ir::*;
-use solve::environment::{Environment, InEnvironment};
 use std::fmt::Debug;
 use std::sync::Arc;
