+++ chalk-rust/src/solve/normalize_application.rs
@@ -1,7 +1,6 @@
 use errors::*;
 use ir::*;
 use solve::Solution;
-use solve::environment::{Environment, InEnvironment};
 use solve::fulfill::Fulfill;
 use solve::infer::InferenceTable;
 use solve::solver::Solver;
